Update Countdown with doubles

+++ src/main/java/utility/countdown/Countdown.java
@@ -16,5 +16,5 @@ public interface Countdown {
      * Returns seconds left
      * @return seconds left
      */
-    int getSecondsLeft();
+    double getSecondsLeft();
 }
